Rewrite source code analyzer
This replaces the original source code analyzer with a more
generalized analyzer.

The original version was more of a "is this even possible?" whereas
this new approach is closer to a real POC.

I tried several approaches that didn't work out so well before
settling on this approach.  This approach will walk the AST
and infer boto3 types based on certain code constructs (importing
`boto3`, accessing `boto.client`, invoking the client creation method,
etc.) and then propagate those changes where possible.  It's
essentially a greatly simplified abstract interpreter where we only
care about tracking boto3 types.

I also integrated the stdlib symtable to do some of the heavy
lifting for tracking global vs. local scope.

The biggest change here is that because of the nature of the
abstract interpreter, code that's never run won't count towards
API calls.  This means it's not enough to define a function, it
actually has to be called somewhere for it to be considered.

To handle chalice views (which are never explicitly called in
the app.py file), a node transformer is used to implicitly
call chalice views functions.