Export the view exceptions into chalice package
There was an inconsistency where you had to use
`chalice.app` to pull in certain exceptions, which was never
the intent.  These are all available in the top level
``chalice`` package now.

Updated docs accordingly.  The old way still works, but the
preferred way is just `from chalice import ForbiddenError`.