Extract APIG method creation into separate class
This is a refactoring that separates the logic for creating
an API gateway method (request method, integration request,
integration response, method response) into a separate class.

With the introduction of CORS support, these calls are no longer
unconditional calls.  This is going to grow even more so in
the future when CORSConfig is introduced along with other optional
values.

By separating this into a separate class, this also gives us the
ability to have much more granular testing when creating the
various methods/integrations, which has the most involved API calls.

As part of this change, I also switched over the
APIGatewayResourceCreator to only use the typed aws client.
This should also make this class easier to test now without requiring
the botocore stubber.