Initial commit of chalice local command
This command will spin up a local http server that you can
use to run your chalice apps.  The intent is to provide a quicker
feedback loop while developing your code.  Previously you would
have to deploy to API gateway in order to test changes to your API.

There's one main feature that's missing from chalice local: this
does not try to assume the role that's associated with lambda function.

In order to do this, we'll likely need to update the trust policy
of the role to give the AWS user being used locally the ability
to assume the role.  Not exactly sure how I want this to work so
I haven't added that as part of this commit.

Closes #22.