Require type information for all chalice code
Type information has been gradually added to
chalice code over time, and as each module is
fully typed, it's been added to the Makefile
to ensure any new code to the module has type
information.

This commit adds type information to all the
remaining code that didn't have types.

As a result, the `typecheck` rule in the Makefile
is now a one liner.  There's a few notes:

* All new code in chalice from this point on will
  need to have type info or else the travis builds
  will fail.
* The analyzer module needed a *lot* of Any types
  to make mypy happy.  The problem is part with
  how the code is structured (fixable) and part
  because of the the type inference mypy seems to
  be deducing from the ast module.