Refactor code analyzer for better type support
This refactors the code analyzer to avoid setting
attributes and returned ast nodes.  Instead the inferred
type is tracked in a separate mapping object that
SymbolTableTypeInfer now depends on.

This allows the type hints to be specific to a given ast node,
so rather than Any they can now be ast.Call, ast.Name, etc.