Bump up initial function timeout to 2.5 minutes
It makes more sense to have a longer intial deploy if it means
you're not going to get an error message about not being
able to create a function because the role we just created isn't
available yet.  I've bumped this up a few times, but setting
it to 2.5 minutes I shouldn't have to mess with this value
in the future.