Pull out policy integration into separate class
This is paying off some technical debt where the lambda
deployer code kept growing as new functionality was being added.

I decided to slice out the policy generation code into a separate
class.  This is the code that bridges the core logic for policy
generation with chalice specific concepts (project dir, app file,
recording/loading last generated policy etc).

This is also nice because we're pulling out more code from
the deployer that's not directly involved with making API calls.

This should also be easier to test (coming in the next commit).