Use proper empty policy when policy.json does not exist
What was happening is that the default policy when
policy.json does not exist was an empty dict.

When that made it's way to the diffing code, which
assumed a precondition that the policies are well formed,
a key error is generated.

To fix this, if the policy.json does not exist for whatever
reason, a default policy is added.

Fixes #29.