Add 'package' command and SAM support
This adds support for a 'package' command.  This command takes the
current application and produces artifacts that can subsequently be
deployed by cloudformation.  The intent is to allow for decoupling the
packaging from the deployment process, which provides a number of
benefits.  First, something besides chalice can do the deployment.  For
many customers that have existing deployment tooling based around
cloudformation, they would be able to leverage these tools to handle
deployment.  In addition to existing customer tooling, this also makes
it possible to integrate with other AWS services.  By creating a
separate app package, chalice can now integrate with AWS CodePipeline.
This allows a user to set up a full CD pipeline that deploys your
chalice app whenver you 'git push' to a repo.

The second benefit is that by decoupling the packaging from the
deployment, this allows a user to modify the app package before
deploying.  While chalice strives to expose APIs that make app
configuration simple, the user can now have a system in place that
packages the app, modifies the app package, and then deploys the app.

Missing pieces:

* Updating the 'deploy' command to be able to deploy an app package.