Add support for authorizer_name in route(...)
I'm proposing making a breaking change and removing
authorizer_id in favor of authorizer_name.

The rationale for this is two-fold.  One, as far as I can tell,
it's impossible to support the old way of using authorizer_id.
This is because the swagger document defines the entire API,
and it appears to (at least conceptually) wipe out the old stuff
before putting in the new.  Therefore it's not possible to refer
to a pre-existing authorizer by id, it has to be called out in
the document.

The second reason is that I think this is the right long term
fix.  I want the chalice app to define all the relevant parts of
the api in code, and being able to define the authorizers in your
app.py and allow chalice to create it for you seems ideal.

Before this change, the workflow was basically 1) Create the authorizer
either via the CLI, console, or API.  2) Find it's auto generated
ID.  3) Specify that authorizer_id in an @app.route(...) call.

Now the workflow is you specify the authorizer definition in your
app file, and your @app.route(...) references that authorizer
by name.