Add chalice stage option and deprecate existing arg
The "chalice deploy <stage>" is deprecated.  Instead, I've
added two options in its place:

* --api-gateway-stage, for users that want the existing behavior
* --stage, for the new stage behavior which creates an entirely separate
  set of AWS resources.

I've also added a deprecation warning if you specify a positional stage
as well as param validation if you specify both the api gateway stage
as well as the positional stage param.

As part of this commit I've also updated the remaining commands that
need to be stage aware:

* package
* url
* generate-sdk
* logs

For each of these commands I've added a "--stage" option that uses the
deployed values JSON file to look up the appropriate stage values.