Remove references to lambda_arn
This is handled by .chalice/deployed.json.
To accomplish this, I had to update the signature of the
API Gateway deployer to pass in the deployed lambda_arn, which
is needed for the swagger document.  I was considering using
the "existing resources" object, but it's not clear if allowing
a partially hydrated deployed resources object is a good idea or not.
For now I'm just explicitly passing in the data I need from
previous deployement steps (just the lambda_arn).