Pull out policy generation into separate object
This is may potentially get more involved once we account for
multiple files.  This object all the logic out of coordinating
with the policy generation code into its own class.
This allows the ApplicationPolicyHandler to focus solely on
whether or not we should be loading policy files (and from where).