Support stage specific policy files
Previously there was only a single .chalice/policy.json file
that you could specify.

Now you can provide a stage specific policy file to control
this value.  The default value for this file is
.chalice/policy-<stage>.json, but to help with migration
the .chalice/policy.json file is still supported for the
dev stage if the .chalice/policy-dev.json file does not
exist.