Add support for environment_variables config option
This builds on the initial work from @emellis and adds
support for the environment variables config option.

These changes include:

* Updating the config object with an environment_variables property
* Adding the concept of "chain merge" to config.  This allows
a dict to be updated (merged) instead of returning on the first value
found.
* Update API based deployer to inject env vars when needed.  This is
needed for both create_function and update_function calls.
* Changed update_function_code to just update_function which will
optionally update function config if required
* Update packager to add environment variables to SAM template.

There's still a few things in the initial PR that I'll need to circle
back on in follow up PRs.