Add validation for python version
Right now this is a soft warning, but we may eventually make this
a hard fail.  Validation happens in two spots:

* When running "chalice deploy"
* When creating a new-project

The `warnings` module is being used because I'm considering making
this a hard fail, in which case I would update this to be a
DeprecationWarning before actually removing this functionality.