Tag deployed lambda functions
Makes it possible for customers to easily see
which lambda functions are used for chalice.
Also opens up the possibility of us not having
to rely on the local deployed.json.  That would
be future work though.  This commit
only adds support for tagging the functions.

The format uses colon separated values because
Lambda does not allow CSV values for the value
of a tag.

I also bumped the min version of botocore to the
version where tagging support was added to lambda.