Mark typing as a third party library
Typing does not exist in the standard library for Python
2, but it is in the standard library for Python 3.5.
Chalice was styled accordling to have typing grouped
with the third party imports as it was written originally in
Python 2. However, this fails linting of import order for
Python 3 and so force typing to always be linted as a third
party import to enforce chalice's original import order pattern.