Increase max parents for a class
This was increased because in python 3, the standard library
`collections.Mapping` class (inherited by
`chalice.app.CaseInsensitiveMapping`) has a depth of 5 and pylint does
not differentiate between chalice classes and standard library classes
in terms of max depth. If there is a way for this linter to ignore
class heirarchy of standard library classes, this value should be
changed back to 5, but unfortunately, it looks like there is no
way currently to ignore stdlib classes.