Add vendor/ to the end of sys.path, not beginning
This issue here is if I have a binary module built specifically
for Lambda (Amazon Linux) and I try to import that module locally
on a non linux machine (e.g a Mac), I'll get import errors because
I'm trying to import a linux binary on a mac.

As a workaround, the vendor/ directory is added to the end of
sys.path so that site-packages in your venv will take precedence.
This allows you to install the Mac binaries into site-packages
via `pip install` and still have the deployment package correctly
pull in the linux specific extension modules.