Fix binary validation logic
The only case we care about is where we might send
base64 content back to the user when they were expecting binary.
The new validation checks if the response is actually binary.
If the response is binary, then the user must specify they
want binary content via the Accept header otherwise we
return an error.