Update deploy interfaces to accept a deployed resources dict
This updates the cases where the deployer accepts a single lambda
ARN.  This is going to have with multi lambda function, so this plumbs
through the change to instead accept a dict.  This gives some
flexibility because we change the dict keys if needed without having
to updated all these interfaces again.

Should make rebasing/shared branches easier.

No function changes.