Added packaging support for whl files.
The algorithm for building a lambda compatible bundle is as follows:

1) Download all packages using pip download without specifying binary
preference. This allows it to download whl files where possible and
sdists otherwise.
2) Try to download lamba-compatible whl files for all sdists
and lambda-incompatible whl files that were downloaded in the first
step. Once this step is completed we have all the whl files are are
going to be able to get from PyPi that are compatible with lambda.
3) Any sdists left over without a corresponding compatible whl file will
now be built to a whl. If this results in an incompatible whl file there
is nothing more we can do to get a compatible one and it is up to the
user to vendor that package themselves.

Now that we have done our best to get a whl file for each dependency, we
do a final pass over the directory and install all compatible whl files
to the lambda package. Any left over dependencies that do not have a
compatible whl file are reported to the user, as they will need to be
manually added to the vendor directory before deployment.

Virtualenv is no longer used so the compat functionality for it is no
longer needed.

Since the packaging process now downloads whl files which vary by python
version the deployment process now rebuilds the deployment package if it
detects that the python version has changed.