Updated download logic and fixed satistifed typecheck, linting.
The download algorithm was changed to download all packages using
whatever it can find rather than forcing --no-binary=:all: which allows
it to download whl files on the first pass. This results in fewer
overlal downloads since now only the whl files that are incompatible
with the lambda runtime, and sdists need to be redownloaded.

Next it will try to download lamba-compatible whl files for all sdists
and lambda-incompatible whl files. Once this step is completed we
have all the whl files are are going to be able to get from PyPi that
are compatible with lambda, any sdists left over without a corresponding
compatible whl file will now be built to a whl.

Now that we have done our best to get a whl file for each dependency, we
do a final pass over the directory and install all compatible whl files
to the lambda package. Any left over dependencies that do not have a
compatible whl file are reported to the user, as they will need to be
manually added to the vendor directory before deployment.

Updated tests. And fixed pip requirement.