Added proper handling for purelib/platlib, use subprocess for pip.
Added tests for installing platlib and purelib from the data directory
of a wheel file's .data directory. Pip uses a subprocess instead of
importing pip and using that for the advantage of not polluting stdout,
stderr with progress bars. More granular error detection for pip usage
was also added. Specifically errors for: timeout, new connection failed,
and not having permission to write to disk. Tests were updated accordingly.