Fix crash when analyzer encounters listcomp/genexprs
Internally, python creates a function for listcomps/genexprs
so we treat it similarly.  There's a new child scope and therefore
a new symbol table we need to fetch.  The only tricky part is that
python names this function 'genexpr' and 'listcomp' even in the
case of multiple genexprs and listcomps.  We're not analyzing the
`generators` attr of the nodes, but we will need to take that
into account when we do.  For now it doesn't matter.