Move env var setting to cli/__init__.py
The env var configuration needs to be set before importing
the app so that any module level code will be able to access
env vars.

We might need to pull this up even higher to the cli() function
because any time we import the chalice app we need to make sure
env vars are set.  For now this is just in the 'local' command.