Fix edge case when building packages with optional c extensions
The edge case occurs when a package can compile optional c speedups
which succeed during the build process but lock the generated wheel file
to the arch the package was built on. This makes the wheel file
incompatible with lambda and causes it to fail to add that dependency
even though it could have by building it without the optional c
extensions.

This is a bit tricky to fix as means we need to selectivly break the
ability to compile c extensions during some runs of our wheel buiding
phase. It is also quite differnet on POSIX and Windows.