Use absolute path for project_dir context object
The --project-dir option was not correctly calculating the path to open up the
config file. This would result in valid `project_dir` values raising an OSError
and presenting this error message to the user:

```
$ ls
drwxr-xr-x  6 chris  staff   204B Jul 12 17:24 sample/
...
...
$ chalice --project-dir sample deploy
Unable to load the project config file. Are you sure this is a chalice project?
```