Change the default api gateway stage name to v1
There's some confusion about chalice stages and api gateway stages
and to make this more clear, I propose changing the default value
for api gateway stages to v1.

This is a backwards compatible change because this value is pulled
from the .chalice/config.json.  This is just changing what the
default value is that's written to the config file as part of the
new-project command.