Fix the cryptography tutorial
The cryptography 2.0 package came out and now has wheels, which
means that the custom vendor example we gave is no longer accurate. I
just pinned the version of cryptography to 1.9 so that the tutorial is
still technically correct. The cryptography package has a wide set of
dependencies that fall into most of the interesting cases, so it is hard
to find a replacement package for the tutorial. Also if we switch to
another package and they support wheels, we will just have to update it
again so I think the best course here is just to version pin.