Fix API Gateway stage not being set correctly
The API Gateway stage is not set correctly since it was renamed from dev
to api. This change enforces the 'api' default all the way through the
config layer and CLI layer rather than making the end code that consumes
the configuration manually enforce a default everywhere the api gateway
stage name is used.

fixes #470