Allow api_gateway_stage to be set from config.json
Anything that has chained precedence can't be defaulted at
the CLI layer because you can't differentiate between
a default value vs. not set.  In this case, `None` is used
as the sentinel to indicate not set.

The defaulting still happens in the CLI factory layer, it's just
that it needs to do in the "defaults" dict that is the final
fallback if no other value for a config key was found in the
various search dicts.

This is how `autogen_policy` is handled, so this change uses
that same approach but for `api_gateway_stage`.

Fixes #495