Poll for a successful 200 after redeploys
Sometimes there's a delay after a redeploy where API
gateway will return 502 responses.  We're not trying to
test that the API is available immediately after a redeploy
(that's beyond our control).  We just want to verify that
after a redeploy, once everything settles, the new routes
we added are available and the routes we removed are no longer
available.