Initial POC of new deployer
The existing deployer module is starting to get out of hand.
Initially it only had to handle API gateway and a single Lambda
function, but over time we've added support for AWS resources,
and continue to add additional resources (#516).

I don't think the existing deployer design is going to hold up
as all these new resources are being added.  The code is hard to
understand, hard to use, and hard to modify.  There's also a number
of reported customer reported bugs from the existing deployer.

Digging further into the deployer code, there are a number of
inconsistencies as well as unexpected and unintuitive behaviors.

I've been experimenting with a new redesign of the deployer to
not only fix the existing issues, but also make future changes
easier.  This initial POC only supports `@app.lambda_function()`
and only sketches out how to do the initial creation of resources,
but I think this should be able to handle the remaining gaps.

To help with initial testing, I've added a temporary `deploy-new`
command so both deployers can live side by side during development.
Before this gets merged to master, `deploy-new` will just become
`deploy` and the old code will be removed.

The `newdeployer` module docstring has a description of how
the new system works, and is a good starting point for a technical
overview.