Add file encoding argument to the get_file_contents
Specifically using the default causes an issue on windows if any file
has contents that are not readable by cp1252. This comes up during the
packaging of a deployment. Some egg-info files contain unicode
characters. Windows fails to build the deployment bundle since it
crashes trying to read the egg-info file.