Fix handling of trailing slash to match API Gateway
API Gateway will not match a URI against a route that has a capture
group as the last path component if that capture group would be filled
with an empty string. Example:

With the route:
/resource/{name}

/resource/bob matches
/resource/    does not match

Previously local mode would match both URIs to that route, setting the
name parameter to an empty string.

closes #582

This change also prvents `chalice local` from running if there is a
route that ends with a / since `chalice deploy` will not let you deploy
such routes.