Switch sphinx doc theme to alabaster
Rationale:

I couldn't get the existing theme to render the sidebar
navigation properly with the level of nesting in our TOC
that we use.  Additionally, this all predates the alabaster
theme being a built in sphinx theme.  There's enough customization
options in alabaster that I can tailor it to exactly what I want.

I had to pull in open-sans so we can use the same fonts used
by the existing theme.