Fix IAM role creation issue
Fixes bug #565.

So there ended up being a couple issue that I fixed here, first being bug #565
which was fixed by making a change to the awsclient.py file.

The second issue that I discovered while working on this bug was the following.
If you do:

```
$ chalice new-project muck
$ cd muck
$ chalice deploy --no-autogen-policy
```

The same bug happens because the policy-env.json file does not exist. This
results in a policy document of:

```
"Version": "2012-10-17",
"Statement": []
```

being uploaded which throws a MalformedPolicyDocument error. In order to fix
this issue I made changes to the deployer.py file. I removed the default policy
and now raise a RuntimeError and alert the user to the fact that they do not
have a policy-env.json file.

The last change that I made was to handle malformed json by adding nicer
logging modeled after the config file loader.