Add delete-new command
The basic idea is to rely on the sweeper code to handle deletions.
Conceptually this is similar to creating an empty chalice app
and running `chalice deploy`.

The other alternative was create a new planner type that specifically
handled deletions.  The reason for leveraging the sweeper for deploys
is two fold:

1. We have to write sweeper logic anyways.  This was originally to
   handle the case where you remove resources from your app.py and
   rerun `chalice deploy`.  This logic has to stay regardless of what's
   decided for deletion
2. A deletion planner would unnecessarily have to take into account
   what's currently specified in your app.py along with the remote state
   (stuff that's already been deployed).  This is unnecessary
   complication.  For deletion we only care about what exists
   remotely.

Another benefit of (2) is that we handle an edge case where you
remove resources from your app.py but don't run `chalice deploy`.
In the old deployer code, we base deletion based on your app.py
so we'd essentially leak a resource.  This is now handled by this
new deletion code because it's always based on the deployed.json
file.