Add support for bz2 compressed sdists from PyPi
Support for bz2 has been removed from PyPi, and is explicitly called out
in PEP 527 as not supported. No new packages can be uploaded with the
bz2 format. However; there are still packages that have not been
migrated over, and possibly never will be. This adds support for bz2
since it is a trivial change and increases the number of packages
Chalice can automatically handle.