Add support for scheduled events in newdeployer
The only significant change from the existing deployer
is that the old deployer treated a scheduled event as
a single resource whereas the new deployer treats it as
two separate resources: a lambda function and a cloudwatch
event.  The main benefit of this approach is that you can
reuse existing logic for building/planning a lambda
function.