Remove role_arn from LambdaFunction model
This was intended as a way to deal with values we didn't know
until a deployment happened.  However, to make the create and update
code paths consistent, this is all done via the deployed values
mechanism as well using a load/store instruction.  In addition
to making creates/updates consistent, it also handles the case where
you can't use an API call to know the appropriate value to use.
An example of this is the rest api id, where the onyl way to know
the correct ID to use is from the deployed.json file.