Update deployed json files to not include stage
Now that deployed values are written out as a separate
file per stage, we no longer need the "stages.<stagename>"
nesting.  It introduces unnecessary ambigous edge cases (what
if there's a "prod" stage key in the dev.json) file?

It also makes it simpler for anything that consumes the deployed
values from the deployer (i.e the reporter).  They don't have to
to a "d['stages'].values()[0]" to figure out the deployed resources.