Create empty deployed resources when no deployment exists
Previously, we returned Optional[DeployedResources], which
made usage awkward because you always had to check if
deployed_resources was not None.  With this change, if there
has been no deployment, we just create an empty DeployedResources.
Consuming code now needs to catch a `ValueError` if they try
to access a resource that was not deployed.