Fix packager edge case normalizing sdist names
If a package name is different from its normalized name then there would
be a packaging issue where it would fail to recognize a valid whl file
it had generated from an sdist. The whl file would be normalized
according to PEP 503 and the sdist one would not causing it not to be
added to the list of valid whl files for deployment.