Fix abi3 packaging
Python 3 has an ABI called abi3 which is the CPython 3 Stable ABI which
is compatible with any version of Python 3 after 3.2. The packager
previously did not recognize these as being compatible with the Lambda
abi of cp36m and would not use them. This fix adds abi3 as a valid abi
for python 3.

This fixes automatic packaging of cryptography 2.2.x since it removed
its specific python 3 wheel files in favor of a single cp34 abi3 wheel.