Fix regression when using role arn in package command
We need a noop method for precreated roles in the template generator.
This was missed in testing because the models were directly
passed to the template generator instead of using the
dependency builder.  In this case we were only passing the
LambdaFunction resource to the template generator, but
in reality we'd also pass the `models.IAMRole` resource
as well because it's a dependency of the lambda function.

Fixes #793.