Upgrade mypy to the latest version
I had to update some of the code to account for the upgrade,
there were a few issues that the new mypy catches.  These were always
errors, just never caught.

The AST changes require casting.  I _think_ these are ok, given
the code is the same as before, we're just now explicitly casting
to a given type to help out mypy.  It's not really possible to
fix this without bigger changes to the analyzer.