Lazy load chalice app in config
This solves the problem of trying to load a chalice app
with the appropriate env vars.  We need to load the config
object in order to know what env vars to set, but the config obj
also imports the app.  To fix this, the config obj can be lazy loaded
by returning a callable.  The config obj will then invoke the
callable to create the chalice app.

I've also shuffled around some of the tests.  I've moved the integ local
tests to be functional.  That way it runs as part of the travis builds.

Builds on #798.
Fixes #748.