Raise validation error on invalid VPC params
This was initially done in chalice.deploy.validate, but it
produces too many false positives.  This way, we only fail if we try
to construct a resource with invalid params.  This does mean we might
have latent issues that we'll only find about when we try to deploy our
app, but that seems better than failing validation when it wouldn't
actually affect a deploy.