Remove models.pyi typings
The latest version of mypy now supports attrs via
its plugin interface.  As a result, we can now put
type definitions directly in the models.py attrs
classes.

In a few places I had to put explicit cast()'s.
These were due to bugs/inconsistencies in the existing
models.pyi.  There are a few things that we can't
express with mypy's typing.  Essentially, there are
some values that are a `Union[A, B]`, but we know for
a fact after certain code paths that the type will
always be `B` and never `A`.  I'm using a cast to
express this.