Add on_sns_message decorator to app
This plumbs through creating the config, and the
event handler and event class.  Right now you just
provide the topic name.  I was considering adding
subscription attributes/filters, but it doesn't look like
SAM supports this so we'd have to drop to cfn for this.
For now we'll just support the topic name.